pyproject.toml
[tool.poetry]
name = "passwordless-login"
version = "0.1.0"
description = ""
authors = ["Mihai Ionut Vilcu <ionutvmi@gmail.com>"]
readme = "README.md"
packages = [
    { include = "app_webui", from = "src" },
    { include = "passwordless", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.0"
uvicorn = { extras = ["standard"], version = "^0.21.1" }
jinja2 = "^3.1.2"
python-dotenv = "^1.0.0"
starsessions = "^2.1.1"


[tool.poetry.group.dev.dependencies]
poethepoet = "^0.19.0"
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
reportUnknownMemberType = false

[tool.poe.tasks]
start = "uvicorn app_webui.main:app --reload"
